Moved the Path extension methods that were in Utils.kt into the internal package
