+++ core/src/main/java/net/corda/core/internal/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * WARNING: This is an internal package and not part of the public API. Do not use anything found here or in any sub-package.
+ */
+package net.corda.core.internal;
