+++ node/src/integration-test/kotlin/net/corda/services/messaging/MQSecurityAsNodeTest.kt
@@ -1,11 +1,11 @@
 package net.corda.services.messaging
 
-import net.corda.core.copyTo
-import net.corda.core.createDirectories
+import net.corda.core.internal.copyTo
+import net.corda.core.internal.createDirectories
 import net.corda.core.crypto.CertificateType
 import net.corda.core.crypto.Crypto
 import net.corda.core.crypto.X509Utilities
-import net.corda.core.exists
+import net.corda.core.internal.exists
 import net.corda.node.utilities.*
 import net.corda.nodeapi.ArtemisMessagingComponent.Companion.NODE_USER
 import net.corda.nodeapi.ArtemisMessagingComponent.Companion.PEER_USER
