+++ node/src/main/kotlin/net/corda/node/internal/NodeStartup.kt
@@ -6,6 +6,8 @@ import joptsimple.OptionException
 import net.corda.core.*
 import net.corda.core.crypto.commonName
 import net.corda.core.crypto.orgName
+import net.corda.core.internal.createDirectories
+import net.corda.core.internal.div
 import net.corda.node.VersionInfo
 import net.corda.core.node.services.ServiceInfo
 import net.corda.core.utilities.Emoji
