+++ node/src/main/kotlin/net/corda/node/services/persistence/NodeAttachmentService.kt
@@ -9,7 +9,7 @@ import com.google.common.io.CountingInputStream
 import net.corda.core.contracts.AbstractAttachment
 import net.corda.core.contracts.Attachment
 import net.corda.core.crypto.SecureHash
-import net.corda.core.isDirectory
+import net.corda.core.internal.isDirectory
 import net.corda.core.node.services.AttachmentStorage
 import net.corda.core.serialization.*
 import net.corda.core.utilities.loggerFor
