+++ node/src/test/kotlin/net/corda/node/services/transactions/PathManagerTests.kt
@@ -1,6 +1,6 @@
 package net.corda.node.services.transactions
 
-import net.corda.core.exists
+import net.corda.core.internal.exists
 import org.junit.Test
 import java.nio.file.Files
 import kotlin.test.assertFailsWith
