+++ test-utils/src/main/kotlin/net/corda/testing/node/MockNode.kt
@@ -11,6 +11,9 @@ import net.corda.core.crypto.cert
 import net.corda.core.crypto.entropyToKeyPair
 import net.corda.core.crypto.random63BitValue
 import net.corda.core.identity.PartyAndCertificate
+import net.corda.core.internal.createDirectories
+import net.corda.core.internal.createDirectory
+import net.corda.core.internal.div
 import net.corda.core.messaging.MessageRecipients
 import net.corda.core.messaging.RPCOps
 import net.corda.core.messaging.SingleMessageRecipient
