+++ test-utils/src/main/kotlin/net/corda/testing/node/NodeBasedTest.kt
@@ -7,6 +7,8 @@ import net.corda.core.*
 import net.corda.core.crypto.X509Utilities
 import net.corda.core.crypto.appendToCommonName
 import net.corda.core.crypto.commonName
+import net.corda.core.internal.createDirectories
+import net.corda.core.internal.div
 import net.corda.core.node.services.ServiceInfo
 import net.corda.core.node.services.ServiceType
 import net.corda.core.utilities.WHITESPACE
