+++ tools/demobench/src/main/kotlin/net/corda/demobench/explorer/Explorer.kt
@@ -1,8 +1,8 @@
 package net.corda.demobench.explorer
 
-import net.corda.core.createDirectories
-import net.corda.core.div
-import net.corda.core.list
+import net.corda.core.internal.createDirectories
+import net.corda.core.internal.div
+import net.corda.core.internal.list
 import net.corda.core.utilities.loggerFor
 import net.corda.demobench.model.JVMConfig
 import net.corda.demobench.model.NodeConfig
