+++ verifier/src/integration-test/kotlin/net/corda/verifier/VerifierDriver.kt
@@ -8,7 +8,7 @@ import com.typesafe.config.Config
 import com.typesafe.config.ConfigFactory
 import net.corda.core.crypto.X509Utilities
 import net.corda.core.crypto.commonName
-import net.corda.core.div
+import net.corda.core.internal.div
 import net.corda.core.map
 import net.corda.core.crypto.random63BitValue
 import net.corda.core.transactions.LedgerTransaction
